requirements.txt
streamlit
openai
gspread
google-auth
streamlit-quill
python-dotenv
reportlab
pandas
beautifulsoup4
pdfkit